Checkpoint -- Somewhat more sane graphing
Add a few more OST graphs (iops, lock rates, client count)
Add server-level graphs in right-hand pane
Fix a fd leak
Organize target-level graphs in collapsible lists

As a proof-of-concept, this has been useful to show us that
RRDtool's public API will probably not get us very far -- the
tricks necessary to avoid writing graphs to disk don't scale.
Running hydra-server with apache+wsgi screws with stdout and
the graphs get eaten/truncated.  Running httpd -X (debugging mode)
is a stupid work-around, but not a solution.

On the javascript side, there seem to be problems with refreshing
large numbers of images serially in a loop.  If all of the visable
graphs fail to refresh before the timeout, the refresh never happens.