Per build versioning, deps, DEBUG = False
Each build will append a date/time stamp to the version so that
we can have some reasonable success at upgrading and keeping up
to date even with fast paced commits.
Add rsyslog as a dependency to force its use.
Consider a build done with make as an indication of "production"
use and disable the DEBUGging.