Assign $(shell ...) vars with simple expanstion
The make assignment operator ?= is a recursive expansion, conditional
on the variable not already being set.  However we want the value of
the date command evaluated only once, not on each use so use simple
expansion (:=) along with the $(origin) function to emulate ?= as a
:= expansion instead of a = expansion.
Also use simple expansion for the VERSION since it's value should
never change during a build.