HYD-285 Client connect/disconnect events
Create events when clients connect and are evicted by the
adminstrator.
This will cause a client connection event storm when a filesystem
comes up.  A possible enhancement in this area is to track client
state and not log events for first-connect-after-filesystem-start
events.
As a bonus, includes an event for an otherwise used lustre port.

Change-Id: I466a0e0110a04273eb4e88e2b0364abb2f40b0d8