HYD-352 Metrics stop working if two creations clash
Using the suggested approach of adding a unique constraint
on the combination of content_type and object_id, we can
prevent the creation of duplicate Database entries.

This does ignore another problem, namely that simply accessing
the Database can trigger its creation if it doesn't already
exist.  This was intended to be a simplification, but should
probably be changed in favor of explicit creation.