HYD-607: Sort out target removal
ManagedTargetMount is no longer a stateful object.  Operations to
set up and tear down the mount locations for a target are done
by ManagedTarget, which has gained an extra state during setup.

This means that for the moment:
 * Removing any host used by a target requires the target to be removed
 * There is no adding or removing secondary mount points from a target
   after creation

In effect both of those were already true, as the existing code wasn't
working properly.  Modifications to the target mounts for a target after
creation will be implemented as Jobs (not StateChangeJobs) later.