HYD-1038: Improve R3D performance using ArchiveRows
This work removes the CDP table in favor of a more column-oriented
approach using a single CDP pickle field for all DS CDPs in a
given row.  Improves intake performance by a factor of 3.  Query
performance was not specifically measured, but should also be
greatly improved, especially when querying large archives.

Change-Id: I2e6c6db8739b7caa8fea99141420633dcd5b8527