Fix broken shutdown/reboot tests
When c604ea26 landed, the tests were not updated to mock
out os._exit(), with the result that running the tests killed
the whole harness. Oops.

Also adds a bit of Makefile goop such that running 'make test'
in the chroma-manager directory also runs 'make test' in the
chroma-agent directory.

Change-Id: I441cd61f41bb53fdcac97d28d62e4b9960900122