HYD-1969: Fix handling of SCSI WWIDs
Previously concatenated serial_80 and serial_83 to get the unique ID for
a ScsiDevice.  This was broken because IRL serial_80 varies for a particular
device depending on the path to the device.

This change is to set a single 'serial' attribute for ScsiDevice, derived
from the agent-reported serial_83 if present and only use serial_80 at all
as a fallback (this order of preference is still customizable in
settings.SERIAL_PREFERENCE)

Change-Id: Ibb6c831972e9ec29962515115d9c38c59a3a7550