HYD-2151 Harden clear_ha() to handle half-configured stonith
The problem here is that pacemaker can get into a state where it wants
to stonith, but because PDUs have not been confgured (they are not
configured for every test) that leave leaves pacemaker with a stonith
agent (fence_chroma) that can't actually do any stonithing.  This is
because the node attributes that tell fence_chroma what to do are
missing due to the lack of PDU configuration.

Where this all becomes a problem is when you try to shut down pacemaker,
it will block the shutdown until it's able to perform the stonith.  So,
let's give pacemaker 30 seconds to shutdown and if it can't let's help
it along.

Once we are sure pacemaker is shut down, if there are any Lustre targets
still mounted, we know we can safely unmount them.

Also, clear devices *after* target stop.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Iab604c0ae6b6949378f85d21fd803d14580f6617