HYD-2087: Improve detection of unknown PDU outlets
The query of a given PDU outlet can sometimes fail, which isn't fatal
but it is annoying. In production, the result would simply be a yellow
outlet status. In test runs, it can cause sporadic test failures due to
test logic expecting outlets to be in known states. This commit improves
the monitoring logic to re-schedule a query job for PDUs with any
outlets in an unknown state.

Change-Id: Ia7c8e945a45e8f6cb0f73f49c1c9d13bf109213d