HYD-2236: Installer broken when opening urls with a user-information part
Urllib2 mistakes the : in the user-information part of an uri for the start of a port identifier.

This patch splits out the user-information part and supplies it as an Auth header.

Authing with HTTPBasicAuthHandler did not send the Auth header so went with this approach: http://stackoverflow.com/questions/2407126/python-urllib2-basic-auth-problem.

Change-Id: I79dd022cff3daeaee47cf9aaf4bcb5581dde8553