HYD-2268: Add validation to session endpoint for POST
This patch adds validation to the session endpoint for post requests. The validation class checks for the existence of the username and password fields.
If that passes, the 403 response has been updated to return a response body if there is an issue authing the user.

This is in the interest of consolidating validation logic on the server and not duplicating it on the clients, and also subscribing to a consistent way of returning validation errors.

Change-Id: I43577aafea4cbc559d2b0ecf55cb9410c86daf27