HYD-2271: Include lustre-client 2.4.0 in IEEL archive
This commit finishes what 152ec2c was supposed to be. Unfortunately,
it wasn't noticed until after landing that some of the changeset
went missing (e.g. creation of lustre-client bundle, apache config)
after it was reworked to pull the special IEEL build of the client.

Creates a lustre-client bundle for the archive. After installation
this bundle is available as an unauthenticated yum repo on the manager
at https://<manager>/client

Change-Id: I6a60760804aec6141ff649b00af0d02f46f27b29