HYD-2194 - Adapt benchmark to find max values
Added a new script which provides a starting SU size, and batch size.
The batch size is to tune the server rego so it can get all the hosts
registered.  Then SU size is the starting size of the tests.  It will
increment after success as this size by 1 SU at a time.

Note:  The incrementation is not reliable.  It's best to kill the script and IML, restart
and try again at the higher number.  I've seen 15 SU's work locally on my mac with this
script.

Change-Id: Ided3aa13a61cd27aafc6c61e91ba1026eafa9b89