HYD-2367 - exercise_filesystem trying to write zero length files.
exercise_filesystem dynamically computes the size of the file to
write to the filesystem as a test to take into account different
disk sizes in different environments. However, _get_stat_simple
could return a zero value if we just don't have the information
yet for a new filesystem, so we need to wait for it to be nonzero.

Change-Id: I9c5adf3da03d8e00419eec0086d9c3012173b8e1