HYD-2429: IE9 uncaught JS errors.
IE was falling back to IE7 standards Document Mode.
Added <meta http-equiv="X-UA-Compatible" content="IE=edge" /> to force IE to latest version.

Also note that port 8000 is forcing the browser mode to IE9 Compatibility, but not much that can be done there as far as I know short of unchecking the setting.

Change-Id: I8cda07a5b84f338cb81d90fc4087758320a1fa6c