HYD-2357: Separate log collection from running tests for trap.
Currently log collection happens in the same script as the
selenium tests are run in. However, we want to be able to
collect the logs even if there is a catastrophic failure
running the tests by putting it into a trap in the test job,
so it needs to be separated out into its own script.

Change-Id: Id2c53f39522117b5be271cc688ec29c89f192df7