HYD-2459: Prevent SSH exceptions from tanking test
When checking to see if a host is shutting down, a SSHException
may be thrown if the check is done at exactly the wrong moment.
This may also happen during sshd startup. In any case, the point
of the test method is to determine if SSH is ready yet, and the
answer is "no".

Change-Id: I06a70a33092e2b9e724bf4c8b2f8c0a32aa0ee61