HYD-2477: Discourage server add with failed validations
When any of the pre-add validations fail, the Confirm button is
disabled and the user has to click a button indicating that they
understand that adding a host without all validations is
unsupported.

Also emits a warning to http.log so that there is evidence that
the validation result was overridden.

Change-Id: If7172b284fe3e8ced16905683338e32b325eb770