HYD-2528: Kill zombie agents
One theme that seems to tie together a number of tickets is the idea
that "there's something wrong with SSL". The basis of this idea seems
to stem from deployments where the agent seems to be running, but not
communicating with the manager. It's not definitive that this would
result from agents that have hung around from previous installs, but
seems to be a plausible hypothesis. In any case, ensuring that we've
got a clean deploy environment by doing the python equivalent of
'killall -9 chroma-agent-daemon' seems reasonable and low-risk.

Change-Id: Ib991ca27748583c23cba6ea54d64c8b8ad0ca999