HYD-1898: incorrect Fail(over|back)TargetJob deps
The dependency set for these jobs included a dependency on the
host being moved away from, which is wrong. We only care about the
host being migrated to. Likely the source of some spurious job
cancellations due to 'lnet_up' jobs racing with hosts booting up
and setting the state on their own.

Change-Id: I9e21c3ffec7b16695c29541a60359297e97f20a4