HYD-2671 Unable to use EMC multipath device with IML Managed Mode
The linux plugin doesn't support directly EMCPower devices and
so doesn't place them as parents of the drives which they contain.

This changes creates that relationship.

There is some discussion about whether this EMCPower support should
be a seperate plugin. I did investigate that and did investigate
making it a subclass of the class's here. My conclusion was this
is not reasonably possible becase the linux device builds relationships
between things like mdraid, emcpower, lvm, and dev mapper. I.e.
If someone builds and LVM on and MDRaid the linux plugin would
detect and report it correctly. For this emcpower plugin to be standalone
it would need to also detect lvm, mapper, mdraid etc so that it could
report the correct relationship between it and those other devices.

It would also mean the Linux plugin was still mis-reporting the
relationship.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Ibf090bdc4da14f03cd2c3ad3a818b6a7e09e49f0