HYD-2430 Server added using password auth can be left in broken state
Improved implementation of determining Jobs/Actions for the case when
the DeployHostJob fails.  In this case, the host is put in the 'deploy_failed' state.
This state will cause an empty list of transitional states to be returned.
The always available AdvertisedJob ForceRemoveHostJob will be available, and the
only choice at this point.
added 'deploy_failed' state for a host
Update GUI in notifications.js to show undeployed icon for deploy_failed.

Change-Id: I1465651245a2a26383e83e52d56cf525b83cded1