HYD-2758: Lustre client management and HSM profile
Add a new server type that is managed but not a Lustre server.
Profiles with a 'worker' attribute can be used for HSM copytool
workers, Lustre client nodes, and any other node type that will
run an IML agent and perform commands, but will not participate
in HA configs or other Lustre server roles.

Note that this functionality is not intended to open the door to
compute client management. Adding that capability to IML would have
to be done very carefully with special consideration given to the
difference in scale between Lustre server count and Lustre compute
client count.

Also adds a new POSIX HSM Worker profile that, when selected,
will provision the server with a Lustre client. The 2.5+ Lustre
client RPMs include a default POSIX copytool implementation. This
profile can be used as the basis for vendor-specific copytools.

Change-Id: Iff7bcac2cc8fe15770640dec88970ed8c68eca9b