HYD-2765: Timeout waiting for outlet disassociation
In ConfigureHostFencingJob.on_success(), there was an erroneous
use of ManagedHost.save(). Instead, we should be using
JobSchedulerClient.notify() to update the host attribute. This bug
was a race, where sometimes the updates would happen OK, and
sometimes the save() would result in an integrity error, killing the
job and never allowing the rest of the jobs in the command to proceed.

Change-Id: I35fe0ff1df9a9923f31516eb2592dfee7fc85d69