build: run decruftinator as an install: dependency
The decruftinator was not reliably running because it
had been set as a dependency of the rpm: target, which
is only used if RPMs are actually built. When cached
RPMs are used, it doesn't get executed. With this fix,
we will reliably check for intra-build cruft and remove
it on every build.

Change-Id: I659cd12195f31afea47571a662429a744ecdf29f