HYD-2851: add debug for test failures
With only one instance of this test failure, it's pretty hard to
say what might have happened. This commit adds an assert to check
that the filesystem says it has more than 0 free bytes before
attempting to write to it. If the test fails again, then it's
likely that the problem is in Lustre itself. Otherwise, there
could be some race that allows the test to start writing on the
client before the filesystem is ready.

Change-Id: I338e6cc1932c36fa69eb4206d12c75cd41a7081c