HYD-2714 HYD-2714 Test failure: No action found with state mounted
Refresh the obj before needing it, to make sure the latest up-to-date
value is available.

Saddly, this fix requires a db hit, and so the test that protects against
unknown db hits creeping in failed.  Glad it failed, but sad that I have to
up the value from 7 to 8.   (Hopefully we'll rearchitect the DB out someday.)

Change-Id: Ib4d24ceb1ce2161f7735a5d018b1af3225d7b494