HYD-2882 Improve canonical representation of devices including multipath
Ensure that a complete cannonical representation table of devices - i.e.
a look up exists for each device path which means if two devices are
actually the same device then the resulting lookup gives the same
output - is complete.

This means that each device detection routing must add itself to the
table and that the table should be recursive to ensure the canonical
value is found.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Ic2379000d5e043abcc089c0baf3ddbd9da7392c8
Reviewed-on: http://review.whamcloud.com/9459
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>