HYD-2714 No action found with state mounted
Moved add_verb from api process to job_scheduler, where
the rest of the stateful transition collecting goes on.

Motivation:  The API process does an RPC call to the JobScheduler to
dynamically fetch the transitions.  Before this patch, the JobScheduler
side would compute a list of transitions based on the state of the object
as it knew it from reading the DB state value.  Then it would send those
transitions back to the API side.  The API side would again check the DB
for the state, and do another state based calculation of annotating UI
verbs to the transitions returned based on the JobClass which is based on the
object state.  The problem is the DB state value can change
while the RPC is in flight.  It is rare, but when it happens it means the
state menu gets the wrong info displayed.  Of course, for the UI and realtime
the user can refresh, or theoretically, the UI would poll again and fix the
problem.

Change-Id: I97d7d99976d8490562748b29a25b2dabb1a91ebd