TEI-847: Can not connect to the ChromeDriver test failure.
Implement workaround until fixed upstream. ChromeDriver has a bug
where it asks for a port from Chrome, but then that port is stolen
by another process before it binds to it, resulting in a failure
to start the ChromeDriver. A ticket has been filed and triaged
upstream, but they don't seem to be in any hurry to fix it, so
I'd like to implement a "try again" workaround on starting the
WebDriver until they fix the bug upstream so we don't lose test
runs to it. This carries no risk in masking a bug in our product
code as this is before our product code comes into play and is
purely an infrastructure issue.

Change-Id: I15a87b691ff930ba6c9ffa73f9626d071538cec7