HYD-3067: Reg/Unreg copytools via JobScheduler
State changes for ObjectCache-managed objects MUST be done
via JobScheduler in order to ensure that the cached instance
stays in synch with the DB.

Without this patch, an obscure (but not impossible) corner case
could lead to the copytool being 'stopped' in the DB but 'started'
in the ObjectCache. This would result in a confusing UI
representation of available actions.

Change-Id: I7c2a36d7cd7c746f703a2afd53091d1f4f826618