HYD-3068 Data inconsistency issue between sessions and plugins
An issue exists where by a session can be used against more than
a single Plugin instance. This causes problems because the mapping
between local and global id's is in the session and so if the
plugin changes this mapping is incorrect or can be incorrect.

This version fixes the bug that highlighted this feature and also
adds an assert to ensure that 1<->1 relationship is kept.

A new ticket will be raised to move some context into the
plugin instance in particular local_id_to_global_id and
global_id_to_local_id as these are mapping the plugin_instance
data they should be with and part of that data.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I3d6617ee71d140653a0ca8037d49f7d126e19cc6
Reviewed-on: http://review.whamcloud.com/9796
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Michael MacDonald <michael.macdonald@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>