HYD-3068 Data insistancy issue between sessions and plugins
An issue exists where by a session can be used against more than
a single Plugin instance. This causes problems because the mapping
between local and global id's is in the session and so if the
plugin changes this mapping is incorrect or can be incorrect.
This version fixes the bug the highlighted this feature and also
adds an assert to ensure that 1<->1 relationship is kept.

The cause of the problem is communication to the plugin's outside
of the scope of a session. Sessions are the protocol and should
always be used for all communication.

This patch fixes the code that demonstrates the problem a follow
on patch will actually be more thorough and make the data
more resilient as well as testing that the condition is
detected.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Ie9963afef31fd1560af8e7f2fd4daa9898a97247