HYD-3021 Expected states can end up with None values
Expected states is the states that an object is expect to be in when
the current set of commands acting on it complete. This information
is derived from locks that exist on the object. However objects can
lock with a target state of None - which means the target state is
unknown.

The line of code that creates expected_states did not take this into
account and so an expected state of None could be computed.

This change stops None being added as an expected state.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I4ea33e3153a24a6bab95d116d0035408593390a7
Reviewed-on: http://review.whamcloud.com/9999
Tested-by: Jenkins
Reviewed-by: Aric Coady <aric.coady@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>