HYD-3050 Database issue during completion_hooks
The problem here is that any job started in a notify
will close the database connection, and so any updates
to the database carried out within that notify are
rollbacked rather than committed.

The removal of the line

django.db.connection.close()

from JobScheduler._spawn_job

This may have knock on effects but there is no way on
earth a function should just close a database connection
it did not open - so at worst we will find out the root
cause of the issue that caused the line to be added in the
first place.

One knock on effect was that commands requested/executed after
a host was deleted caused problems. This is as expected the code
was carrying out Fence Configuration to a host that was now
deleted. A small change to detect this was added, a more general
solution might be required but for the 2.1 release this is
effective.

Bug fix in notify which means that the command is correctly
reused if another job is added.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Id92719aeee1902457d6247b9ec4fa501e60d12b5
Reviewed-on: http://review.whamcloud.com/10037
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>