HYD-3140 HSM excludes linux plugin but linux plugin contains shared PluginAgentResources resource
HSM excludes linux plugin but linux plugin contains shared PluginAgentResources
resource this causes some asserts when trying to communicate with the Worker plugin.

The right solution is to move PluginAgentResources into a plugin of it's own. But
given the current closeness to release, a better solution is to make the worker
agent load the linux plugin normally and then have the linux plugin not report
resources on worker nodes.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I0e809cff56d485fff57992a83b7db5757e425f9b
Reviewed-on: http://review.whamcloud.com/10051
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Michael MacDonald <michael.macdonald@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>