HYD-3141 Refactor inheritors of AgentRxQueue to remove class variables and needless classes
AgentRxQueue was inherited in a strange way when a simple call to the initiator to
AgentRxQueue would do.

For example

{code}
class CorosyncRxQueue(AgentRxQueue):
    plugin = COROSYNC_PLUGIN_NAME

...

    self._queue = CorosyncRxQueue()
{code}

can be replaced with

{code}
    self._queue = AgentRxQueue(COROSYNC_PLUGIN_NAME)
{code}

There is no advantage to the child classes, these changes were done in HYD-3068 but
this ticket splits them out into a separate ticket for landing.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I77ed0e04dc32658a85bf13b88b4aab6c6e03bc11
Reviewed-on: http://review.whamcloud.com/10053
Tested-by: Jenkins
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>