HYD-3096: Cannot start filesystem using simulator
This fixes an issue creating a mgs key to store as state key for the simulator.
Previously, we sorted the outer list first then the inner lists. If the items in the
inner list had inconsistent orders before sorting, the key could have different results.

This patch sorts the inner lists first, then the outer list to ensure a repeatable sort.

Change-Id: I0d6c4fa789823fd7fd3a49c8b4311f1a805f48c7
Reviewed-on: http://review.whamcloud.com/10112
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>