TEI-2040: Use a trap to gather upgrade test report files.
The IML Upgrade test works a bit differently than our
other tests as it doesn't use "|| true" when running
the tests so that they will halt on failure, but these
halts mean that the line that scp's the test reports
doesn't get executed if the tests fail, rather than
gathering the test report about the failure. The
solution to this is to move the gathering of the xml
reports into a trap so that it will execute whenever
the test script that follows terminates.

Change-Id: Ia5a9b7ce07db866e95b2c336098a4a7a9766be0d
Reviewed-on: http://review.whamcloud.com/10131
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>