HYD-2826 Update boilerplate.py to make updating copyrights for 2014.
This is probably an over the top change to make it possible to update
the Copyright message in the boilerplate.

The change is needed to allow the old boiler plates to be varied, to
make it so that the boiler plate is in a file which will be useful for
the javascript in the future (says Joe).

This code now has minimal language specifics, has any number of old
boilerplates. The old and new boilerplates are kept in files in a
directory boilerplates under this script.

I also refactored a lot to try and give it some shape that in the future
might be useful. It's an for discussion, this code is not part of
the shipping product.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I1a669caf0c6693dfbeee87b1a4f00d2fff4f5b04
Reviewed-on: http://review.whamcloud.com/10148
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>