HYD-3209 LNet stalemate issue
This patch is the first in a series of patches for this issue. It
changes the gui and rest interface to use/provide a member_of_active_filesystem
element to the host.

The field is calculated by checking against all active filesystems.

In future this flag will be replaced with a state called active which
will allow all the dry run mechanisms to work correctly.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I14c8dfeddb6bacbe2aceed3889d80a81b2fd1849
Reviewed-on: http://review.whamcloud.com/10302
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>