HYD-3209 LNet stalemate issue
This change implements the backend for this change.

It also caches the new lnet configuration so that the cache version
can be used when lnet is not up. This was done previously but this
change extends it so that changes to the configuration are cached for
cases where lnet configuration is changed whilst lnet is not up.

To do this some of the values return by lnet read have been removed,
this is because there is no way of reproducing them at present and
at present they are not used.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I48a23fab06b2518b52971b883e9af349054dd1fe
Reviewed-on: http://review.whamcloud.com/10307
Tested-by: Jenkins
Reviewed-by: Aric Coady <aric.coady@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/10523