HYD-3227: Include SLES RPMs in IEEL tarball
This commit adds plumbing to automatically include client/server
bundles for SLES11sp3. The SLES RPMs are pulled from the same job
as the EL6 RPMs. Note that these bundles are included in the
installer, but they are not listed in any profile and therefore
are not registered with the manager during installation.

Refactors the build goop a bit to DRY up common information
such as jenkins credentials, root URL, and lustre build job/number.

Fixes a bug in find_deps.py by simplifying the logic for looping
over bundle packages.

Change-Id: If418a13db5594e7f0d722ab43c359b2256bb14b0
Reviewed-on: http://review.whamcloud.com/10330
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/10381
Reviewed-by: Joe Grund <joe.grund@intel.com>