HYD-3210 - Test failure: SSI, LoadLNetStep, Communications error
Test:  SSI, test_firewall_access, test_agent

Stop the debugging/logging from failing by indicating to remote_command 
that the return value should not be checked.  Because checking causes 
an error to be raised but that then masks the error that was raised 
previously.  This means that the output of remote_command is always 
presented to the debug stream.

Change-Id: I243e87cce104d29642a4cacf864f2eb77bba5740
Reviewed-on: http://review.whamcloud.com/10355
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>