HYD-3124 Chroma agent always install pacemaker even in management mode
Thhis change remove some dependancies on pacemaker from the chroma-agent
repo and also updates the install so that the managed host and monitored host
installation installs different yum packages.
This is done by passing an gent_managed parameter to the code api to
fetch slightly different deploy python.

There is also a requirement to not load the corosync plugin and some minor
references to management code that had to be removed.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Id57c8d453772b02191bb8bb4132be6a8d267f5eb
Reviewed-on: http://review.whamcloud.com/10184
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/10397