HYD-3235 Filesystem detection not consistantly using normalized device name
The existing file system detect scan was using a normalized path but not
a normalized block id. So whils the device it used to detect was correct
the block id information it filled in was not correct. For example it
would us /dev/mapper/mpathX to detect the existing filesystem but then
use the block id from the underlying device such as /dev/sdX. Due to valid
caching the underlying device may not have been updated and may contain
something quite different.

Another issue fixed and associated with the bug is that the complete
normalization tree for paths was not being created. This has been improved.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I592f11efbb8855c8cb54799aa0027b2cd45282ff
Reviewed-on: http://review.whamcloud.com/10466
Tested-by: Jenkins
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/10522
Tested-by: Chroma Test User <hpdd-devops@intel.com>