HYD-3264: suggested changes Kombu-3.0.19 + Celery-3.1.10 upgrade
RpcTimeout is related to the fact that internal queues are not binded
to rabbitmq queues anymore.

Few "RuntimeError('Acquire on closed pool')" errors being raised and
we were seeing lots of: "connection_closed_abruptly" in the rabbitmq
logs. It has been suggested by kombu developers to update kombu
and its dependencies.


Change-Id: If94dee7eb13497ebd9cbe649dcd8f6fa9c0b921e
Reviewed-on: http://review.whamcloud.com/10681
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>