HYD-2483 Return all metrics if no metric names supplied.
Currently, one has to know which stats one wants to retrieve
when querying the metrics API. This is fine for the JS client, 
since we already know that stuff.

The idea is that if no metric names are supplied, then all
metrics for the object (or the list of object if it's a 
list resource) are retrieved.

Change-Id: I656b7866597913b163bc655932a741615d31c379
Reviewed-on: http://review.whamcloud.com/10666
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>