HYD-3317 V2.1.0RC2 and later do not install manage_updates and so future attempts at updates fail.
A packaging issue means that the manage_updates.py file is part of
the chroma-agent-management rpm and not the chroma-agent rpm. This
means that whilst you can install to RC[2-5] but will have trouble
upgrading from RC[2-5]. Interestingly if you upgraded to RC[2-5]
you will probably be alright because the manage-updates file will
already be there.

Durring testing another couple of update related qwerks where discovered
and fixed. The host would try to start corosync and do reboots
on a monitored system and also when part of a running filesystem
monitored hosts could not be upgraded. This is obviously
not correct as upgrade does not effect the running filesystem.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I54ec732d849c7d2a0ae3a4c0bfbd3a33e8d73636
Reviewed-on: http://review.whamcloud.com/10952
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>