HYD-3290 mdraid detection fails due to normalization issues
This is a generic issue in that detection always uses normalized
device names and in fact this code was improved to make it more
robust, however device detection is much more vague in the way
it describes devices. Storing the name /dev/md/xxxxx when the
normalized name is /dev/disk/by-id/yyyy-yyy-yyy-yy-y this change
toughens up the way in which mdraid and emcpower devices are
reported making sure that they are normalized to the same
standard used by the detection routines.

Change-Id: Id6c128de92f999268038719f7e4586795907e680
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/10777
Tested-by: Jenkins
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>