HYD-3290 mdraid detection fails due to normalization issues
This is a generic issue in that detection always uses normalized
device names and in fact this code was improved to make it more
robust, however device detection is much more vague in the way
it describes devices. Storing the name /dev/md/xxxxx when the
normalized name is /dev/disk/by-id/yyyy-yyy-yyy-yy-y this change
toughens up the way in which mdraid and emcpower devices are
reported making sure that they are normalized to the same
standard used by the detection routines.

This patch is exactly as per the master except that it keeps
the catch all exception in linux_network.py. This is kept for
safety on the 2.0.1 release

Change-Id: Ie90ac3b9c7014f988079f3a97f0a315aa5f93209
Chang-Id: Id6c128de92f999268038719f7e4586795907e680
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/10950
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>