HYD-3327 Backend changes to make cleanup after failed install cleaner.
This change cleans up the host installation in a number of ways.

Removal of the deploy_failed state, this state is really the same
as undeployed and doesn't add any functionality.

When a host fails to deploy it will stay in the undeployed state.

The ability to deploy is now possible from two usages of the host api

1. The existing creation of a host using create operation

2. By setting the state of an existing host that is in the undeployed state

In the latter case the information provided as part of the create operation
must be supplied in the state change operation.

The information required can be assertained by looking at the new
property of a host called install_method this can take fource values.

manual, id_password_root, private_key_choice and existing_keys_choice

The put method of redeployment cannot be used if manual the install method.

If ssh_key is the install method then all that is required is for the
state to be set to something valid such as configure.

If ssh_pw is ithe install method then root_pw, pkey, pkey_pw need to be
supplied in the same format as for a create.

Change-Id: I2f65b8b71b4111fa2dbecb4fd50b323a098515ff
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/10853
Tested-by: Jenkins
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>